+++ src/soopervisor/commons/__init__.py
@@ -1,11 +1,12 @@
 from soopervisor.commons import conda, docker, source, dependencies
-from soopervisor.commons.dag import load_tasks, find_spec
+from soopervisor.commons.dag import load_tasks, load_dag, find_spec
 
 __all__ = [
     'conda',
     'docker',
     'source',
     'load_tasks',
+    'load_dag',
     'find_spec',
     'dependencies',
 ]