+++ src/soopervisor/commons/dependencies.py
@@ -7,9 +7,9 @@ def check_lock_files_exist():
     if not Path('environment.lock.yml').exists() and not Path(
             'requirements.lock.txt').exists():
         raise click.ClickException("""
-Expected requirements.txt.lock or environment.lock.yml at the root directory, \
+Expected requirements.lock.txt or environment.lock.yml at the root directory, \
 add one and try again.
 
-pip: pip freeze > requirements.txt.lock
+pip: pip freeze > requirements.lock.txt
 conda: conda env export --no-build --file environment.lock.yml
 """)