+++ tests/airflow/test_airflow_export.py
@@ -4,13 +4,16 @@ import subprocess
 import importlib
 from unittest.mock import Mock, ANY
 from pathlib import Path
+import json
 
 from airflow import DAG
-from airflow.providers.docker.operators.docker import DockerOperator
-from ploomber.io import _commander, _commander_tester
+from airflow.providers.cncf.kubernetes.operators.kubernetes_pod \
+    import KubernetesPodOperator
 import pytest
 
+from conftest import _mock_docker_calls
 from soopervisor.airflow.export import AirflowExporter, commons
+from soopervisor.exceptions import ConfigurationError
 
 
 def git_init():
@@ -21,26 +24,12 @@ def git_init():
     subprocess.check_call(['git', 'commit', '-m', 'commit'])
 
 
-def _mock_docker_calls(monkeypatch, cmd, proj):
-    tester = _commander_tester.CommanderTester(
-        return_value={
-            ('docker', 'run', f'{proj}:latest', 'python', '-c', cmd): b'True\n'
-        })
-
-    subprocess_mock = Mock()
-    subprocess_mock.check_call.side_effect = tester
-    subprocess_mock.check_output.side_effect = tester
-    monkeypatch.setattr(_commander, 'subprocess', subprocess_mock)
-
-    return tester
-
-
 @pytest.fixture
 def mock_docker_calls(monkeypatch):
     cmd = ('from ploomber.spec import '
            'DAGSpec; print("File" in '
            'DAGSpec("pipeline.yaml").to_dag().clients)')
-    yield _mock_docker_calls(monkeypatch, cmd, 'sample_project')
+    yield _mock_docker_calls(monkeypatch, cmd, 'sample_project', 'latest')
 
 
 @pytest.fixture
@@ -48,7 +37,7 @@ def mock_docker_calls_serve(monkeypatch):
     cmd = ('from ploomber.spec import '
            'DAGSpec; print("File" in '
            'DAGSpec("pipeline.serve.yaml").to_dag().clients)')
-    yield _mock_docker_calls(monkeypatch, cmd, 'sample_project')
+    yield _mock_docker_calls(monkeypatch, cmd, 'sample_project', 'latest')
 
 
 @pytest.fixture
@@ -56,7 +45,7 @@ def mock_docker_calls_callables(monkeypatch):
     cmd = ('from ploomber.spec import '
            'DAGSpec; print("File" in '
            'DAGSpec("pipeline.yaml").to_dag().clients)')
-    yield _mock_docker_calls(monkeypatch, cmd, 'callables')
+    yield _mock_docker_calls(monkeypatch, cmd, 'callables', 'latest')
 
 
 # need to modify the env.airflow.yaml name
@@ -104,7 +93,8 @@ def test_airflow_add_sample_project(monkeypatch, tmp_sample_project,
 
 def test_airflow_export_sample_project(monkeypatch, mock_docker_calls,
                                        tmp_sample_project,
-                                       no_sys_modules_cache):
+                                       no_sys_modules_cache,
+                                       skip_repo_validation):
     load_tasks_mock = Mock(wraps=commons.load_tasks)
     monkeypatch.setattr(commons, 'load_tasks', load_tasks_mock)
 
@@ -126,13 +116,21 @@ def test_airflow_export_sample_project(monkeypatch, mock_docker_calls,
                                             mode='incremental')
     assert isinstance(dag, DAG)
     assert set(dag.task_dict) == {'clean', 'plot', 'raw'}
-    assert set(type(t) for t in dag.tasks) == {DockerOperator}
+    assert set(type(t) for t in dag.tasks) == {KubernetesPodOperator}
     assert {n: t.upstream_task_ids
             for n, t in dag.task_dict.items()} == {
                 'raw': set(),
                 'clean': {'raw'},
                 'plot': {'clean'}
             }
+
+    spec = json.loads(Path('serve', 'sample_project.json').read_text())
+
+    assert [t['command'] for t in spec['tasks']] == [
+        'ploomber task raw --entry-point pipeline.yaml',
+        'ploomber task clean --entry-point pipeline.yaml',
+        'ploomber task plot --entry-point pipeline.yaml'
+    ]
 
 
 @pytest.mark.parametrize('mode, args', [
@@ -142,8 +140,8 @@ def test_airflow_export_sample_project(monkeypatch, mock_docker_calls,
 ],
                          ids=['incremental', 'regular', 'force'])
 def test_export_airflow_callables(monkeypatch, mock_docker_calls_callables,
-                                  tmp_callables, no_sys_modules_cache, mode,
-                                  args):
+                                  tmp_callables, no_sys_modules_cache,
+                                  skip_repo_validation, mode, args):
     exporter = AirflowExporter(path_to_config='soopervisor.yaml',
                                env_name='serve')
 
@@ -162,7 +160,7 @@ def test_export_airflow_callables(monkeypatch, mock_docker_calls_callables,
     # check tasks in dag
     assert set(dag.task_dict) == {'features', 'fit', 'get', 'join'}
     # check task's class
-    assert set(type(t) for t in dag.tasks) == {DockerOperator}
+    assert set(type(t) for t in dag.tasks) == {KubernetesPodOperator}
     # check dependencies
     assert {n: t.upstream_task_ids
             for n, t in dag.task_dict.items()} == {
@@ -175,10 +173,13 @@ def test_export_airflow_callables(monkeypatch, mock_docker_calls_callables,
     # check generated scripts
     td = dag.task_dict
     template = 'ploomber task {} --entry-point pipeline.yaml'
-    assert td['get'].command == template.format('get') + args
-    assert td['features'].command == template.format('features') + args
-    assert td['fit'].command == template.format('fit') + args
-    assert td['join'].command == template.format('join') + args
+    assert td['get'].arguments == [template.format('get') + args]
+    assert td['features'].arguments == [template.format('features') + args]
+    assert td['fit'].arguments == [template.format('fit') + args]
+    assert td['join'].arguments == [template.format('join') + args]
+
+    assert {t.image for t in td.values()} == {'image_target:latest'}
+    assert {tuple(t.cmds) for t in td.values()} == {('bash', '-cx')}
 
 
 def test_stops_if_no_tasks(monkeypatch, mock_docker_calls, tmp_sample_project,
@@ -198,8 +199,23 @@ def test_stops_if_no_tasks(monkeypatch, mock_docker_calls, tmp_sample_project,
     assert 'has no tasks to submit.' in captured.out
 
 
+def test_validates_repository(monkeypatch, mock_docker_calls,
+                              tmp_sample_project, no_sys_modules_cache):
+    exporter = AirflowExporter(path_to_config='soopervisor.yaml',
+                               env_name='serve')
+
+    exporter.add()
+
+    with pytest.raises(ConfigurationError) as excinfo:
+        exporter.export(mode='incremental')
+
+    assert str(
+        excinfo.value) == ("Invalid repository 'your-repository/name' "
+                           "in soopervisor.yaml, please add a valid value.")
+
+
 def test_skip_tests(monkeypatch, mock_docker_calls, tmp_sample_project,
-                    no_sys_modules_cache, capsys):
+                    no_sys_modules_cache, skip_repo_validation, capsys):
     exporter = AirflowExporter(path_to_config='soopervisor.yaml',
                                env_name='serve')
 
@@ -215,7 +231,8 @@ def test_skip_tests(monkeypatch, mock_docker_calls, tmp_sample_project,
 
 # TODO: check with packaged project
 def test_checks_the_right_spec(monkeypatch, mock_docker_calls_serve,
-                               tmp_sample_project, no_sys_modules_cache):
+                               tmp_sample_project, no_sys_modules_cache,
+                               skip_repo_validation):
     shutil.copy('pipeline.yaml', 'pipeline.serve.yaml')
 
     exporter = AirflowExporter(path_to_config='soopervisor.yaml',