+++ tests/kubeflow/test_kubeflow_config.py
@@ -1,0 +1,10 @@
+from soopervisor.kubeflow.config import KubeflowConfig
+
+
+def test_default_values(session_sample_project):
+    config = KubeflowConfig.defaults()
+
+    assert config == {
+        'backend': 'kubeflow',
+        'repository': 'your-repository/name'
+    }