+++ tutorials/airflow/ml-intermediate.py
@@ -1,23 +1,29 @@
 import json
 from pathlib import Path
 
+from kubernetes.client import models as k8s
+
 from airflow import DAG
 from airflow.utils.dates import days_ago
 from airflow.providers.cncf.kubernetes.operators.kubernetes_pod \
     import KubernetesPodOperator
+
+volume = k8s.V1Volume(name='shared-folder', host_path=dict(path='/host'))
+volume_mount = k8s.V1VolumeMount(mount_path='/mnt/shared-folder',
+                                 name='shared-folder')
 
 default_args = {
     'start_date': days_ago(0),
 }
 
 dag = DAG(
-    dag_id='{{project_name}}',
+    dag_id='ml-intermediate',
     default_args=default_args,
-    description='Ploomber DAG ({{project_name}})',
+    description='Ploomber DAG (ml-intermediate)',
     schedule_interval=None,
 )
 
-path_to_spec = Path(__file__).parent / '{{project_name}}.json'
+path_to_spec = Path(__file__).parent / 'ml-intermediate.json'
 spec = json.loads(path_to_spec.read_text())
 
 for task in spec['tasks']:
@@ -27,14 +33,13 @@ for task in spec['tasks']:
         arguments=[task['command']],
         dag=dag,
         task_id=task['name'],
-        name='{{project_name}}-pod',
-        # if no File.client is configured, you may use volumes and
-        # volume_mounts to share data among pods. See the tutorial in our
-        # documentation for an example
+        name='ml-intermediate-pod',
+        volumes=[volume],
+        volume_mounts=[volume_mount],
         in_cluster=False,
         do_xcom_push=False,
-        # if pulling a local image, set image_pull_policy='Never'
-    )
+        # to force local image pulling
+        image_pull_policy='Never')
 
 for task in spec['tasks']:
     t = dag.get_task(task['name'])