+++ CHANGELOG.rst
@@ -1,8 +1,22 @@
 CHANGELOG
 =========
 
-0.5.2dev
+0.6.1dev
 --------
+
+0.6 (2022-01-04)
+----------------
+* Adds support for SLURM
+* ``AirflowExporter`` uses ``KubernetesPodOperator`` by default (#33)
+* Simplified Airflow and Argo/k8s tutorials
+
+0.5.2 (2022-01-02)
+------------------
+* Clearer error message when pending ``git commit``
+* Clearer error message when the user does not change docker repository default value (#29)
+* Argo spec sets ``imagePullPolicy`` to ``Never`` if repository is ``null``
+* Documents Kubernetes/Argo configuration schema
+* General documentation improvements
 
 0.5.1 (2021-07-26)
 ------------------